Start implementing login logic
Only account 2000000's requests will be handle by this server.
Indeed, i have figured out that it won't be possible, at least not easy,
to implement feature partially, i.e: implement some feature on this
server and proxy other features to hercules because it would imply to
have a synchronized state between this server and hercules.
For example if i want to implement login for this server and bypass
login server on hercules, it won't be possible(without changes) as char server expect to
have a state created by login server.

The idea is to implement features here, and test them with account
2000000.
Hecules Map server state will be replicated on this server at the end.
At least i expect this server to have a more complete state of the game
than hercules map server.
Client connected with Account 2000000 will see more thing than client
connected with account 2000001